--- src/main/java/my/project/domain/Symbol.java
@@ -1,18 +1,0 @@
-package my.project.domain;
-
-public enum Symbol {
-    VERTICAL("|"),
-    BRIDGE("-----"),
-    NONE("     ");
-
-    private final String symbol;
-
-    Symbol(String symbol) {
-        this.symbol = symbol;
-    }
-
-    public String getSymbol() {
-        return symbol;
-    }
-
-}
