+++ src/main/java/my/project/step4/domain/Direction.java
@@ -1,0 +1,8 @@
+package my.project.step4.domain;
+
+/**
+ * Created : 2020-12-11 오전 11:23
+ * Developer : Seo
+ */
+public class Direction {
+}
