+++ src/main/java/my/project/step4/domain/LadderLine.java
@@ -1,0 +1,10 @@
+package my.project.step4.domain;
+
+
+/**
+ * Created : 2020-12-11 오전 11:17
+ * Developer : Seo
+ */
+public class LadderLine {
+
+}
