+++ src/main/java/my/project/step4/domain/Point.java
@@ -1,0 +1,8 @@
+package my.project.step4.domain;
+
+/**
+ * Created : 2020-12-11 오전 11:21
+ * Developer : Seo
+ */
+public class Point {
+}
