+++ src/test/java/my/project/step4/domain/DirectionTest.java
@@ -1,0 +1,9 @@
+package my.project.step4.domain;
+
+
+/**
+ * Created : 2020-12-11 오전 11:24
+ * Developer : Seo
+ */
+public class DirectionTest {
+}
