+++ src/test/java/my/project/step4/domain/LadderLineTest.java
@@ -1,0 +1,9 @@
+package my.project.step4.domain;
+
+
+/**
+ * Created : 2020-12-11 오전 11:21
+ * Developer : Seo
+ */
+class LadderLineTest {
+}
