+++ src/test/java/my/project/step4/domain/PointTest.java
@@ -1,0 +1,8 @@
+package my.project.step4.domain;
+
+/**
+ * Created : 2020-12-11 오전 11:22
+ * Developer : Seo
+ */
+public class PointTest {
+}
